move stat keeper class
