+++ tests/golem/task/test_taskcomputer.py
@@ -5,10 +5,9 @@ from mock import MagicMock, Mock
 
 from golem.client import ClientTaskComputerEventListener
 from golem.task.taskbase import ComputeTaskDef
-from golem.task.taskcomputer import TaskComputer, PyTaskThread, StatsKeeper
+from golem.task.taskcomputer import TaskComputer, PyTaskThread
 from golem.tools.assertlogs import LogTestCase
 from golem.tools.testdirfixture import TestDirFixture
-from golem.tools.testwithdatabase import TestWithDatabase
 
 
 class TestTaskComputer(TestDirFixture, LogTestCase):
@@ -228,35 +227,3 @@ class TestTaskThread(TestDirFixture):
         self.assertGreater(tt.end_time - tt.start_time, 0)
         self.assertLess(tt.end_time - tt.start_time, 20)
         self.assertTrue(tc.counting_task)
-
-
-class TestStatsKeeper(TestWithDatabase):
-    @staticmethod
-    def _compare_stats(stat_keeper, stats):
-        assert [stat_keeper.global_stats.computed_tasks,
-                stat_keeper.global_stats.tasks_with_timeout,
-                stat_keeper.global_stats.tasks_with_errors,
-                stat_keeper.session_stats.computed_tasks,
-                stat_keeper.session_stats.tasks_with_timeout,
-                stat_keeper.session_stats.tasks_with_errors] == stats
-
-    def test_stats_keeper(self):
-        st = StatsKeeper()
-        assert isinstance(st, StatsKeeper)
-        self._compare_stats(st, [0] * 6)
-
-        st.increase_stat("computed_tasks")
-        self._compare_stats(st, [1, 0, 0] * 2)
-        st.increase_stat("computed_tasks")
-        self._compare_stats(st, [2, 0, 0] * 2)
-        st.increase_stat("computed_tasks")
-        self._compare_stats(st, [3, 0, 0] * 2)
-
-        st2 = StatsKeeper()
-        self._compare_stats(st2, [3] + [0] * 5)
-        st2.increase_stat("computed_tasks")
-        self._compare_stats(st2, [4, 0, 0, 1, 0, 0])
-        st2.increase_stat("computed_tasks")
-        self._compare_stats(st2, [5, 0, 0, 2, 0, 0])
-        st.increase_stat("computed_tasks")
-        self._compare_stats(st, [6, 0, 0, 4, 0, 0])
